+++ app/extend/response.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const getType = require('mime-types').contentType;
+const jsonpBody = require('jsonp-body');
 const isJSON = require('koa-is-json');
+const REAL_STATUS = Symbol('Context#realStatus');
 
 module.exports = {
   set length(n) {
@@ -42,5 +44,53 @@ module.exports = {
     const type = this.get('content-type');
     if (!type) return '';
     return type.split(';')[0];
+  },
+
+  /**
+   * set jsonp response body
+   * If client requests with `query[options.callback]`, it will return jsonp body
+   * otherwise it will return json body
+   *
+   * Notice: you can't read `response.jsonp`, you can only get by `response.body`
+   *
+   * @member {Void} Response#jsonp
+   * @param {Object} obj response object
+   */
+  set jsonp(obj) {
+    const options = this.app.config.jsonp;
+    const jsonpFunction = this.ctx.query[options.callback];
+    if (!jsonpFunction) {
+      this.body = obj;
+    } else {
+      this.set('x-content-type-options', 'nosniff');
+      this.type = 'js';
+      this.body = jsonpBody(obj, jsonpFunction, options);
+    }
+  },
+
+  /**
+   * read response real status code.
+   *
+   * e.g.: Using 302 status redirect to the global error page
+   * instead of show current 500 status page.
+   * And access log should save 500 not 302,
+   * then the `realStatus` can help us find out the real status code.
+   * @member {Number} Context#realStatus
+   */
+  get realStatus() {
+    if (this[REAL_STATUS]) {
+      return this[REAL_STATUS];
+    }
+    return this.status;
+  },
+
+  /**
+   * set response real status code.
+   *
+   * @member {Void} Response#realStatus
+   * @param {Number} status the real status code
+   */
+  set realStatus(status) {
+    this[REAL_STATUS] = status;
   },
 };