+++ examples/cookie/app/controller/forget.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = function* () {
-  this.deleteCookie('remember');
-  this.redirect('/');
-};