+++ examples/middleware/config/config.default.js
@@ -1,9 +1,0 @@
-'use strict';
-
-exports.middleware = [
-  'hello',
-];
-
-exports.hello = {
-  text: 'Hello World',
-};