+++ examples/restful_api/config/config.default.js
@@ -1,5 +1,0 @@
-'use strict';
-
-exports.rest = {
-  urlprefix: '/api/',
-};