+++ test/fixtures/apps/cors/config/plugin.js
@@ -1,3 +1,0 @@
-'use strict';
-
-exports.cors = true;