+++ test/fixtures/apps/demo/app/controller/hello.js
@@ -1,6 +1,6 @@
 'use strict';
 
 module.exports = function*() {
-  this.setCookie('hi', 'foo');
+  this.cookies.set('hi', 'foo');
   this.body = 'hello';
 };