+++ test/fixtures/apps/demo/config/config.default.js
@@ -1,2 +1,2 @@
 exports.keys = 'foo';
-exports.protocolHeaders = 'X-Forwarded-Proto';
+exports.proxy = true;